Files are now sectioned out properly with proper imports etc
